[FE-#64] Generalize routes handling for frontend apps
