+++ apps/landing-unl-seminar-v1/src/app/app-routing.module.ts
@@ -1,5 +1,8 @@
+// Core
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule } from '@angular/router';
+
+// Routes
 import { appRoutes } from './app.routes';
 
 @NgModule({
